+++ AppFairApp.xcconfig
@@ -8,10 +8,10 @@ PRODUCT_NAME = Stanza Redux
 PRODUCT_BUNDLE_IDENTIFIER = app.Stanza-Redux
 
 // The semantic version for the app matching the git tag for the release
-MARKETING_VERSION = 0.2.7
+MARKETING_VERSION = 0.2.8
 
 // The monotonically increasing number specifying the internal app version
-CURRENT_PROJECT_VERSION = 109
+CURRENT_PROJECT_VERSION = 110
 
 // The category of the app should be one of: public.app-category.business, developer-tools, education, entertainment, finance, games, graphics-design, healthcare-fitness, lifestyle, medical, music, news, photography, productivity, reference, social-networking, sports, travel, utilities, video, weather
 APP_CATEGORY = public.app-category.reference
